BYO
This is a software to control and manage real time data. In this case
is bio-data using a Myo armband https://www.myo.com/
The software allows 3 types of visualisations and 3 sonic interactions.
It also allows to send data via OSC and save the session as a TXT or
XML file.
Please, note that the TXT file can reach a very large size and the XML
file has a size limit.
The code is supposed to work with any device that can stream data. It
just need to be readapted substituting the Myo values with the
respective values.